refactor: Host 校验抽为独立中间件 host-guard，index.ts 只留组合
